Move co-occurrence rule tests into XmlLanguageSerializationTests fixture
Per review feedback, consolidate the dangling co-occurrence rule tests
into the existing XmlLanguageSerializationTests fixture instead of a
separate test class.
